refactor(kolme): extract request envelope module ownership (#1958)
